use i64 to avoid conversion into i32

+++ src-tauri/src/api/graphql/objects.rs
@@ -21,7 +21,7 @@ pub(in crate::api) struct Semester {
 
     /// The year of the semester.\
     /// _Example_: `2021`
-    year: i32,
+    year: i64,
 
     /// The start date of the semester.\
     /// _Example_: `2021-09-01`
